Added Colors to draw

+++ src/ColorTableClass.hpp
@@ -42,7 +42,7 @@ public:
 	~ColorTableClass();
 
 	void PrintColors(void);
-	int GetColorForKey(std::string key);
+	uint16_t GetColorForKey(std::string key);
 	void setNew888Color(std::string name, uint8_t red,uint8_t green,uint8_t blue);
 	void setNew555Color(std::string name, uint8_t red,uint8_t green,uint8_t blue);
